Move RuntimeConfig into caliper/ src dir
