+++ src/caliper/Caliper.cpp
@@ -12,11 +12,11 @@
 
 #include "Blackboard.h"
 #include "MetadataTree.h"
+#include "RuntimeConfig.h"
 
 #include "caliper/common/Node.h"
 #include "caliper/common/Log.h"
 
-#include "../common/RuntimeConfig.h"
 
 #include "../services/Services.h"
