+++ src/caliper/ChannelController.cpp
@@ -7,7 +7,7 @@
 
 #include "caliper/common/Log.h"
 
-#include "../common/RuntimeConfig.h"
+#include "RuntimeConfig.h"
 
 using namespace cali;
