+++ src/caliper/RuntimeConfig.cpp
@@ -5,7 +5,7 @@
 
 #include "RuntimeConfig.h"
 
-#include "util/parse_util.hpp"
+#include "../common/util/parse_util.hpp"
 
 #include <algorithm>
 #include <cassert>
