+++ src/caliper/config_sanity_check.cpp
@@ -5,7 +5,7 @@
 
 #include "caliper/common/Log.h"
 
-#include "../common/RuntimeConfig.h"
+#include "RuntimeConfig.h"
 
 #include <algorithm>
 #include <iterator>
