+++ src/caliper/test/test_postprocess_snapshot.cpp
@@ -2,7 +2,7 @@
 
 #include "caliper/Caliper.h"
 
-#include "../../common/RuntimeConfig.h"
+#include "../RuntimeConfig.h"
 
 #include <gtest/gtest.h>
