+++ src/common/Log.cpp
@@ -6,11 +6,9 @@
 
 #include "caliper/common/Log.h"
 
-#include "RuntimeConfig.h"
-
+#include <cstdlib>
 #include <cstring>
-#include <memory>
-#include <map>
+#include <string>
 
 using namespace cali;
 
@@ -18,49 +16,44 @@ struct LogImpl {
     // --- data
 
     static const char* s_prefix;
-    static const char* s_spec;
 
     static LogImpl* s_instance;
 
     enum class Stream { StdOut, StdErr, None, File };
 
-    ConfigSet m_config;
-
     Stream        m_stream;
     std::ofstream m_ofstream;
     int           m_verbosity;
-
-    std::string m_prefix;
+    std::string   m_prefix;
 
     // --- helpers
 
     void init_stream(const std::string& name)
     {
-        const std::map<std::string, Stream> strmap { { "none", Stream::None },
-                                           { "stdout", Stream::StdOut },
-                                           { "stderr", Stream::StdErr } };
-
-        auto it = strmap.find(name);
-
-        if (it == strmap.end()) {
+        if (name == "stderr")
+            m_stream = Stream::StdErr;
+        else if (name == "stdout")
+            m_stream = Stream::StdOut;
+        else if (name == "none")
+            m_stream = Stream::None;
+        else {
             m_stream = Stream::File;
-
             m_ofstream.open(name);
 
             if (!m_ofstream && m_verbosity > 0)
                 std::cerr << s_prefix << "Could not open log file " << name << std::endl;
-        } else
-            m_stream = it->second;
+        }
     }
 
     // --- interface
 
-    LogImpl() : m_prefix { s_prefix }
+    LogImpl() : m_verbosity { 0 }, m_prefix { s_prefix }
     {
-        ConfigSet config = RuntimeConfig::get_default_config().from_spec(s_spec);
+        char* val = getenv("CALI_LOG_VERBOSITY");
+        m_verbosity = val ? std::max(0, std::stoi(val)) : 0;
 
-        m_verbosity = config.get("verbosity").to_int();
-        init_stream(config.get("logfile").to_string());
+        val = getenv("CALI_LOG_LOGFILE");
+        init_stream(val ? val : "stderr");
     }
 
     std::ostream& get_stream()
@@ -77,25 +70,6 @@ struct LogImpl {
 };
 
 const char* LogImpl::s_prefix = "== CALIPER: ";
-const char* LogImpl::s_spec = R"json(
-{
- "name": "log",
- "config":
- [
-  {
-   "name": "verbosity",
-   "type": "uint",
-   "value": "0",
-   "description": "Verbosity level"
-  },{
-   "name": "logfile",
-   "type": "string",
-   "value": "stderr",
-   "description": "Log file name or stdout|stderr"
-  }
- ]
-}
-)json";
 
 LogImpl* LogImpl::s_instance = nullptr;
