+++ src/services/Services.cpp
@@ -9,8 +9,6 @@
 
 #include "caliper/common/Log.h"
 
-#include "../common/RuntimeConfig.h"
-
 #include <cassert>
 #include <cctype>
 #include <map>
