+++ test/cali-annotation-perftest.cpp
@@ -27,7 +27,7 @@
 
 #include <caliper/cali.h>
 
-#include "../src/common/RuntimeConfig.h"
+#include "../src/caliper/RuntimeConfig.h"
 
 #include "../src/tools/util/Args.h"
